requirements.txt
cryptography==41.0.1
dnspython==2.3.0
Faker==18.10.0
pymongo==4.3.3
python-dotenv==1.0.0
boto3==1.26.106
python-dotenv==1.0.0
ecdsa==0.18.0